requirements.txt
cohere>=5.0.0
chromadb>=0.4.22
sentence-transformers>=2.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
python-multipart>=0.0.6
langdetect>=1.0.9
PyPDF2>=3.0.0
